+++ src/main/scala/i4s/opencv/core/model/Scalar4i.scala
@@ -3,8 +3,8 @@ package i4s.opencv.core.model
 object Scalar4i {
   def apply(s: org.bytedeco.opencv.opencv_core.Scalar4i): Scalar4i = new Scalar4i(s.get(0),s.get(1),s.get(2),s.get(3))
 
-  def apply(values: Byte*): Scalar4i = {
-    val vs: Array[Byte] = values.toArray.take(4).padTo(4, 0)
+  def apply(values: Int*): Scalar4i = {
+    val vs: Array[Int] = values.toArray.take(4).padTo(4, 0)
     new Scalar4i(vs(0), vs(1), vs(2), vs(3))
   }
